refactor: move return type extensions to extensions dir
