+++ src/Extensions/Types/ReturnTypes/AuthManagerExtension.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace NunoMaduro\Larastan\ReturnTypes;
+namespace NunoMaduro\Larastan\Extensions\Types\ReturnTypes;
 
 use Illuminate\Auth\AuthManager;
 use NunoMaduro\Larastan\Concerns;
