+++ src/Extensions/Types/ReturnTypes/EloquentBuilderExtension.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace NunoMaduro\Larastan\ReturnTypes;
+namespace NunoMaduro\Larastan\Extensions\Types\ReturnTypes;
 
 use Illuminate\Database\Eloquent\Builder as EloquentBuilder;
 use Illuminate\Database\Eloquent\Collection;
