+++ src/Extensions/Types/ReturnTypes/Helpers/AppExtension.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace NunoMaduro\Larastan\ReturnTypes\Helpers;
+namespace NunoMaduro\Larastan\Extensions\Types\ReturnTypes\Helpers;
 
 use Illuminate\Foundation\Application;
 use NunoMaduro\Larastan\Concerns\HasContainer;
