+++ src/Extensions/Types/ReturnTypes/Helpers/RedirectExtension.php
@@ -2,27 +2,27 @@
 
 declare(strict_types=1);
 
-namespace NunoMaduro\Larastan\ReturnTypes\Helpers;
+namespace NunoMaduro\Larastan\Extensions\Types\ReturnTypes\Helpers;
 
+use function count;
 use PhpParser\Node\Expr\FuncCall;
 use PHPStan\Analyser\Scope;
 use PHPStan\Reflection\FunctionReflection;
 use PHPStan\Type\DynamicFunctionReturnTypeExtension;
 use PHPStan\Type\ObjectType;
-use PHPStan\Type\StringType;
 use PHPStan\Type\Type;
 
 /**
  * @internal
  */
-final class UrlExtension implements DynamicFunctionReturnTypeExtension
+final class RedirectExtension implements DynamicFunctionReturnTypeExtension
 {
     /**
      * {@inheritdoc}
      */
     public function isFunctionSupported(FunctionReflection $functionReflection): bool
     {
-        return $functionReflection->getName() === 'url';
+        return $functionReflection->getName() === 'redirect';
     }
 
     /**
@@ -33,11 +33,10 @@ final class UrlExtension implements DynamicFunctionReturnTypeExtension
         FuncCall $functionCall,
         Scope $scope
     ): Type {
-        // No path provided, so it returns a UrlGenerator instance
         if (count($functionCall->args) === 0) {
-            return new ObjectType(\Illuminate\Contracts\Routing\UrlGenerator::class);
+            return new ObjectType(\Illuminate\Routing\Redirector::class);
         }
 
-        return new StringType();
+        return new ObjectType(\Illuminate\Http\RedirectResponse::class);
     }
 }
