+++ src/Extensions/Types/ReturnTypes/Helpers/RequestExtension.php
@@ -2,24 +2,30 @@
 
 declare(strict_types=1);
 
-namespace NunoMaduro\Larastan\ReturnTypes\Helpers;
+namespace NunoMaduro\Larastan\Extensions\Types\ReturnTypes\Helpers;
 
+use function count;
+use PhpParser\Node\Expr\Array_;
 use PhpParser\Node\Expr\FuncCall;
 use PHPStan\Analyser\Scope;
 use PHPStan\Reflection\FunctionReflection;
+use PHPStan\Type\ArrayType;
 use PHPStan\Type\DynamicFunctionReturnTypeExtension;
 use PHPStan\Type\MixedType;
 use PHPStan\Type\ObjectType;
 use PHPStan\Type\Type;
 
-class TransExtension implements DynamicFunctionReturnTypeExtension
+/**
+ * @internal
+ */
+final class RequestExtension implements DynamicFunctionReturnTypeExtension
 {
     /**
      * {@inheritdoc}
      */
     public function isFunctionSupported(FunctionReflection $functionReflection): bool
     {
-        return $functionReflection->getName() === 'trans';
+        return $functionReflection->getName() === 'request';
     }
 
     /**
@@ -30,9 +36,13 @@ class TransExtension implements DynamicFunctionReturnTypeExtension
         FuncCall $functionCall,
         Scope $scope
     ): Type {
-        // No path provided, so it returns a Translator instance
+        // No arguments, it returns a request object from the container
         if (count($functionCall->args) === 0) {
-            return new ObjectType(\Illuminate\Contracts\Translation\Translator::class);
+            return new ObjectType(\Illuminate\Http\Request::class);
+        }
+
+        if (isset($functionCall->args[0]->value) && $functionCall->args[0]->value instanceof Array_) {
+            return new ArrayType(new MixedType(), new MixedType());
         }
 
         return new MixedType();
