+++ src/Extensions/Types/ReturnTypes/Helpers/UrlExtension.php
@@ -2,27 +2,27 @@
 
 declare(strict_types=1);
 
-namespace NunoMaduro\Larastan\ReturnTypes\Helpers;
+namespace NunoMaduro\Larastan\Extensions\Types\ReturnTypes\Helpers;
 
-use function count;
 use PhpParser\Node\Expr\FuncCall;
 use PHPStan\Analyser\Scope;
 use PHPStan\Reflection\FunctionReflection;
 use PHPStan\Type\DynamicFunctionReturnTypeExtension;
 use PHPStan\Type\ObjectType;
+use PHPStan\Type\StringType;
 use PHPStan\Type\Type;
 
 /**
  * @internal
  */
-final class ResponseExtension implements DynamicFunctionReturnTypeExtension
+final class UrlExtension implements DynamicFunctionReturnTypeExtension
 {
     /**
      * {@inheritdoc}
      */
     public function isFunctionSupported(FunctionReflection $functionReflection): bool
     {
-        return $functionReflection->getName() === 'response';
+        return $functionReflection->getName() === 'url';
     }
 
     /**
@@ -33,10 +33,11 @@ final class ResponseExtension implements DynamicFunctionReturnTypeExtension
         FuncCall $functionCall,
         Scope $scope
     ): Type {
+        // No path provided, so it returns a UrlGenerator instance
         if (count($functionCall->args) === 0) {
-            return new ObjectType(\Illuminate\Contracts\Routing\ResponseFactory::class);
+            return new ObjectType(\Illuminate\Contracts\Routing\UrlGenerator::class);
         }
 
-        return new ObjectType(\Illuminate\Http\Response::class);
+        return new StringType();
     }
 }
